+++ lib/runners/processor/gometalinter.rb
@@ -33,7 +33,7 @@ module Runners
     def setup
       add_warning_for_deprecated_linter(alternative: analyzers.name(:golangci_lint),
                                         ref: "https://github.com/alecthomas/gometalinter/issues/590",
-                                        deadline: Time.new(2020, 4, 30))
+                                        deadline: Time.new(2020, 5, 31))
 
       with_import_path do |path|
         if path